kernel: mtk_eth_soc: use ADMAv1 on MT7981 and MT7986
ADMAv2 hangs when receiving corrupted and/or truncated Ethernet
frames. Use ADMAv1 instead on MT7981 and MT7986 which are affected by
the problem which is what MediaTek did as well in their SDK.

Doing so costs about 10% additional CPU load when receiving in
userspace eg. when running iperf3.

However, as there is currently no good way to detect the RX DMA hang
condition that is still the best option.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>